Moved unit test package, updated access modifiers
